Firmware snapshot for npu-1.32.0 drop
